chore: remove unused api hook
